fix(deals): four defects the review found in the reversal path
The past-date test read the SESSION's date, not the installation's. A close
date is a DATE, so which day it falls on is a question about the reporting zone
— asked in the session zone, a deal near local midnight is judged against the
wrong day and the provisional flag goes the wrong way. It now takes the same
installationToday reading rejectPastCloseDate does, which is the rule this path
is the narrow exception to.

The provisional flag skipped its conflict check whenever the restored date was
past, and was then forced true. A person who CONFIRMED the machine's date
cleared that flag deliberately, so an undo could overwrite their answer without
asking. The check now runs for that field too; only the value it would restore
is overridden.

The conflict check ran before the already-restored check, so a person who put a
field back by hand had their undo refused for having already done it. Asked in
the other order, their work is recognised as the undo's own outcome.

An undo could be reapplied through the approval door. close_date_correction is
in AutoApplyKinds, so a rep with autonomy on has staged confirms redeemed
unattended, and a card staged before a reversal and redeemed after it would put
back what the reversal removed. The version pin already refused this — a
reversal writes the deal, so the pinned compare loses — but the pin is a
property of how the KIND is staged, and a memory that holds only while a
configuration elsewhere stays put is not one to rely on. The effect now consults
reversal memory directly.

Also: CorrectionForAudit took only the object grant, so a caller with deal:read
and no row scope could learn a deal exists from a lookup keyed on an audit id;
it now takes EnsureVisible. And RevertCorrection asked "already taken back"
before checking authority, leaking a distinct refusal for a deal the caller
cannot see; the probe now comes first.
